+++ src/pages/List/Header/ListPageHeader.scss
@@ -13,14 +13,14 @@
 
   .listPageHeaderTxt {
     height: 80px;
-    color: white;
+    color: #ffffff;
     font-size: 68px;
     font-weight: 700;
   }
 
   .listPageHeaderTxt2 {
     font-size: 26px;
-    color: white;
+    color: #ffffff;
     letter-spacing: 0.2em;
   }
 }
