+++ src/pages/List/Nav/ListNavCategory.scss
@@ -17,7 +17,7 @@
         text-decoration: none;
 
         &:hover {
-          color: black;
+          color: #000000;
         }
       }
     }
