+++ src/pages/List/Nav/ListNavTitle.scss
@@ -1,7 +1,7 @@
 .listNavTitle {
   display: flex;
   justify-content: space-between;
-  border-bottom: black solid 2px;
+  border-bottom: #000000 solid 2px;
   width: 100%;
   padding-bottom: 15px;
 
@@ -13,7 +13,7 @@
     width: 110px;
     height: 32px;
     border: none;
-    background-color: white;
+    background-color: #ffffff;
     font-size: 14px;
     outline: transparent;
   }
